requirements.txt
streamlit==0.73.1
pandas==1.0.1
Pillow==8.1.0
